mvp/requirements.txt
requests
pydantic
pydantic-settings
python-dotenv